requirements.txt
Werkzeug==2.0.0
dash==2.0.0
numpy==1.21.3
pandas==1.3.4
Pillow==9.2.0
plotly==5.6.0
scikit_learn==1.1.1
